+++ tests/CsvTest.php
@@ -2,6 +2,7 @@
 
 namespace LeagueTest\Csv;
 
+use League\Csv\BOM;
 use League\Csv\Exception\InvalidArgumentException;
 use League\Csv\Exception\RuntimeException;
 use League\Csv\Reader;
@@ -102,15 +103,15 @@ class CsvTest extends TestCase
     public function testBOMSettings()
     {
         $this->assertSame('', $this->csv->getOutputBOM());
-        $this->csv->setOutputBOM(Reader::BOM_UTF8);
-        $this->assertSame(Reader::BOM_UTF8, $this->csv->getOutputBOM());
+        $this->csv->setOutputBOM(BOM::UTF8);
+        $this->assertSame(BOM::UTF8, $this->csv->getOutputBOM());
         $this->csv->setOutputBOM('');
         $this->assertSame('', $this->csv->getOutputBOM());
     }
 
     public function testAddBOMSequences()
     {
-        $this->csv->setOutputBOM(Reader::BOM_UTF8);
+        $this->csv->setOutputBOM(BOM::UTF8);
         $expected = chr(239).chr(187).chr(191).'john,doe,john.doe@example.com'.PHP_EOL
             .'jane,doe,jane.doe@example.com'.PHP_EOL;
         $this->assertSame($expected, (string) $this->csv);
@@ -121,16 +122,16 @@ class CsvTest extends TestCase
         $expected = 'john,doe,john.doe@example.com'.PHP_EOL
             .'jane,doe,jane.doe@example.com'.PHP_EOL;
         $reader = Reader::createFromString($expected);
-        $this->assertNotContains(Reader::BOM_UTF8, (string) $reader);
+        $this->assertNotContains(BOM::UTF8, (string) $reader);
     }
 
     public function testChangingBOMOnOutput()
     {
         $text = 'john,doe,john.doe@example.com'.PHP_EOL
             .'jane,doe,jane.doe@example.com'.PHP_EOL;
-        $reader = Reader::createFromString(Reader::BOM_UTF32_BE.$text);
-        $reader->setOutputBOM(Reader::BOM_UTF8);
-        $this->assertSame(Reader::BOM_UTF8.$text, (string) $reader);
+        $reader = Reader::createFromString(BOM::UTF32_BE.$text);
+        $reader->setOutputBOM(BOM::UTF8);
+        $this->assertSame(BOM::UTF8.$text, (string) $reader);
     }
 
     public function testDetectDelimiterList()