+++ tests/StreamIteratorTest.php
@@ -2,6 +2,7 @@
 
 namespace LeagueTest\Csv;
 
+use League\Csv\BOM;
 use League\Csv\Exception\InvalidArgumentException;
 use League\Csv\Reader;
 use League\Csv\Statement;
@@ -82,13 +83,13 @@ class StreamIteratorTest extends TestCase
     {
         return [
             'withBOM' => [[
-                [Reader::BOM_UTF16_LE.'john', 'doe', 'john.doe@example.com'],
+                [BOM::UTF16_LE.'john', 'doe', 'john.doe@example.com'],
                 ['jane', 'doe', 'jane.doe@example.com'],
             ], 'john'],
             'withDoubleBOM' =>  [[
-                [Reader::BOM_UTF16_LE.Reader::BOM_UTF16_LE.'john', 'doe', 'john.doe@example.com'],
+                [BOM::UTF16_LE.BOM::UTF16_LE.'john', 'doe', 'john.doe@example.com'],
                 ['jane', 'doe', 'jane.doe@example.com'],
-            ], Reader::BOM_UTF16_LE.'john'],
+            ], BOM::UTF16_LE.'john'],
             'withoutBOM' => [[
                 ['john', 'doe', 'john.doe@example.com'],
                 ['jane', 'doe', 'jane.doe@example.com'],